requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
comtypes==1.1.10
Django==3.2.5
djangorestframework==3.12.4
idna==3.2
pyttsx3==2.90
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
waitress==2.0.0
whitenoise==5.3.0
